refactor: add crate "base"
