+++ crates/base/src/lib.rs
@@ -1,0 +1,7 @@
+#![feature(core_intrinsics)]
+
+pub mod error;
+pub mod scalar;
+pub mod search;
+pub mod sys;
+pub mod vector;
