--- crates/service/src/prelude/scalar/mod.rs
@@ -1,7 +1,0 @@
-mod f16;
-mod f32;
-mod sparse_f32;
-
-pub use f16::F16;
-pub use f32::F32;
-pub use sparse_f32::{SparseF32, SparseF32Ref};
